Don't translate email footer
When maybe_mail() is called we are not guaranteed to be using the
recipient's language in order to correctly translate the footer.
Rather than sending out an email with a footer in an unexpected
language, always send it out in English -- so at least we're
consistently wrong when we're wrong.

Task 1898