Use unique URLs for "proofread next page"
In the Enhanced proofreading interface, when proofreaders use the
"Start Proofreading" link, then the "Save and Proofread Next" button,
and then use the browser's back button they can be presented with an
image that does not match the text. If they then save the page the
incorrect project page will be updated.

To fix this, we use a redirect from within the proofreading frame
to ensure that "Start Proofreading" or "Save & Proofread Next" lead
to a URL that is unique to that page. If the user tries to nagivate
back using the browser, they will be presented with an error because
the page state does not match the expected state.

Task 1911