Update F1 access email
Per request from the GM, clarify the pages-per-day portion of the
email that is sent automatically to new F1 formatters.

At @cpeel's suggestion, updated so that the daily page limit for
F1 at the time the email is sent will be inserted automatically,
since we can now store a daily page limit for each round.