move color CSS from layout.less to theme.less
This moves almost everything that can take a color out of layout.less
and into theme.less. This includes border*, color and background-color.
After experimenting with just breaking out the border color, I came
to the conclusion that it was better to move the whole border property
instead of breaking it up.  There are still a few border: none; and
background-color: transparent; declarations that should probably be
moved. This should make it easier to build additional site themes.

Instead of moving the stragglers, we might want to consider merging
layout.less into theme.less: the problem with the current setup is
that many structures have declaration blocks in both files, which can
be confusing.