Improve regex to create URLs into links
Replace the regex used to create URLs into links in task comments
with one that doesn't pull in some trailing punctuation. While these
are perfectly valid in URLs, they are often not intended as part of
the URL. I expect this will cause fewer incorrect links than the prior
version.