Add new charcoal theme
This just themes the basic site pages -- Front Page, Activity Hub,
Round and Pool pages, Project Pages, and the like.

Items deferred:
  1. Proofreading interface and related -- that's a separate process
  2. Special Day colors -- hopefully will be resolved before the
     final MR.
  3. Logo and graphs -- needs more investigation. After we drop
     support for IE-11, we should be able to use SVGs directly, but
     we'll still need to figure out how.
  4. Where Project Managers force background colors in PCs without
     specifying text color, there may be some readability issues.

Some theming work was necessary in page_interfaces so that the colors
for text areas buttons and menus would match the proofing interface
instead of the theme (specifically in the Standard PI and WC).