update-ppv-report: updated content
Removed function tr_w_one_cell($content), as it was not used anywhere
in the script.

The initial commit combined "General Information" and "Difficulty
Details" into "Project Difficulty", changed "Project Difficulty" to
"Post-Processing Difficulty".

Added more links to the form and a couple to the preview page. Changed
"Submit" button to "Preview".