background-colors: standardize bg color names
This commit standardizes the common background color names, more or
less (there are still some oddballs), and tries to use names that
will give some indication of usage.

Also included, an adjustment to the defined general purpose tables:
for the striped tables, flip odd and even so that the even uses the
page background and the odd uses the stripe color, and give all
striped tables a header color. This should not affect the round
listings.