text-border-colors: naming border & text colors
Most of the color naming has been done already, but globally used
borders were deferred until the end, as well as similar considerations
for text colors.

There were still a few spots in the code that had inline color styles
that needed to be moved out. Cleaning that up is also part of this
branch.