Use language script as part of spellcheck
When presented with text containing two different scripts, aspell
behaves differently based on the dictionary. Words containing non-Latin
characters are ignored completely with the English dictionary, yet
words containing Latin characters are flagged as spelling errors with
the Ancient Greek dictionary.

To fix this, restrict misspelled words (as returned from aspell) to
only words that contain the script of the language we're checking.
This ensures that only words with Latin characters are returned when
checking against the English or French dictionary, and only Greek
words are returned when using the Ancient Greek dictionary.