Properly escape special regex chars in filter
When building the regex used to enforce character suites, we need to
correctly escape special regex characters. All of the special regex
characters are in the Basic Latin set and that character suite is
defined in a way where this will never happen for it. However, PMs
can include them in project custom characters and can cause the
filtering to fail.