Use a POST for the actual file load
Use a POST for the actual file load instead of a GET and disable the
button after a user has clicked on it. This helps prevent users from
hitting the link twice inadvertantly and causing a race condition where
two processes are both trying to load at the same time, one of them
failing and reporting failures to the user.