Add users.email index and drop redundant index
For the users table, add an index on email and drop the redundant
UNIQUE index on username (username is the primary key so it is
unique by definition).

In making these changes it became apparent that 8ec37f41f missed
updating the initial DB schema when adding the api_key fields, so
this fixes that too.