Update Proofing Interface Help
Initial purpose of this MR was to remove reference to a button that
no longer is used, but it turned out that more needed fixing than
that, and this turned into a major update to `faq/prooffacehelp.php`.
The purpose of this commit is to remove what was wrong and make an
attempt to clarify some of the text.