Add max per page limit to list_etexts.php
The star text lists can be changed via the URL to set the number of
items to display per page. Set the max to display per page to 100.
This is an arbitrary max, and can be changed if appropriate, but
leaving it with no max has the potential of someone trying to display
a list of everything on one page.