Fix the PP alert threshold
When we switched from using projects.modifieddate to using
user_project_info.t_latest_home_visit to determine when a PPer
was notified of projects needing to be renewed, we left out the
case where a project was reserved in the rounds and checked out
automatically -- t_latest_home_visit is not updated when a project
moves into a post-processing stage. This changes the query to use
the more recent date for determining when it's time to renew.

I've verified on the command line that the query works, but can't
think of a good way to verify it otherwise, so I haven't set up a
sandbox for it.