Fixes sort order of mentor projects.
Change sort order of lists of projects at the top of the for_mentors
script to display projects available and waiting based on when they
left the mentored round, and include the date and time that each
project left the mentored round so that mentors can see for themseves
when each project left the mentored round.

Existing sort is based on `modifieddate`, which may or may not give
the correct information on which project slice left the mentored
round soonest; because of the duplicate author block, if there
are a lot of different titles, slices can be released out of order.