Add the ability to disable special days
Until now once a special day was created and enabled so that
Project Managers could assign the special day to a project, it
was permanent. Fix it so that a special day can be disabled for
assigning to projects, but can remain for projects that used it
prior to it being disabled. Also allow disabled special days
to be selected in the project search form, and add an indicator
to the `show_specials` list to indicate which special days have
been disabled.