Fix snapshot table display
There are a couple of rows in the PP/PPV section of the Site Progress
Snapshot table that extend beyond that section of the table if the
user has filters enabled. Removed IE hack and reworked the styling to
avoid extending the visible row where inappropriate. Also removed
`if` condition used to block filling out the remainder of the row
above the filter row for non-round/-pool stages.