Prevent editing word lists for archived projects
Word lists are stored in the project directory and thus if the
directory has been deleted (because the project was archived or
deleted) the project's word lists cannot be managed.

Prevent the link from showing up on the project page and do a more
robust job of checking for these edgecases in the edit word lists
page and failing early.