Don't include URLs in credits for PG uploads
GM request; PG has asked that URLs (including email addresses) not be
included in uploads to PG. Investigation showed that some of the
image sources' credit lines included links. This MR updates the edit
form for PMs to edit the project pages, and the Image Sources
management form and display page to explicitly emphasize that the
Credits line should not include URLs.