Update wording on AH and shapshot legend
Both the tooltip for ⓘ on the Activity Hub and the description in
site_progress_snapshot_legend.php say "are" eligible. This commit
changes that to "may be" eligible (for example, most people are not
eligible to work as PPVers, yet everyone sees that icon for the PPV
activity. Add an explanation to the Activity descriptions for PP
and SR that the SR count is included in the "in PP" count.