Updated quiz numbering, moderate and basicx
GM request: the quiz page numbering system for the actual quiz and
tutorial pages for the moderate quiz and tutorial pages differs from
that used on the quiz start page, and the file names. This has been
known to cause confusion when one person uses one numbering system,
and another uses the other. This updates the numbering used by those
tutorial and quiz pages to match those on the quiz start page.