Import dp_branding into top level directory
The dp_branding directory is outside the `/c` directory structure.
Add it to the pgdp-production branch as a local hack so that it is
available on both PROD and TEST, and can be updated in both places
as part of the code deployment process and not have someone have to
remember to make sure to update both places when updates are made.