Make word list trimming more efficient
Rather than trim wordlist files via callback after loading the file
into an array, trim the files as a string before exploding it. This
offloads the work to preg_replace() and avoids a callback on every
single word with a tradeoff of needing to load the entire file into
memory.