Cache valid transitions for subsequent calls
Building out the state update form for authorized users on the PM and
Search pages is expensive because we need to loop through _all_ possible
transitions to weed them out. Build and use a cache that narrows the
valid transitions from a given state and use that in subsequent calls
for projects in the same state.