Allow selecting non-enabled data from endpoints
The API was written for the "enabled" filter to be used as a boolean
flag to either return all items or just enabled items. The Swagger UI
expects it to accept both a "true" and "false" value as a boolean which
is confusing for users.

Update the 3 endpoints that accept an "enabled" flag to work both ways.
This enables them to work with the Swagger UI as well as the original
design.