Enable bypassing error log on DB query failure
There are rare times when we know a query might fail and we don't want
to log an error to the error log. Create a temporary table is one of them
because we can't otherwise determine if the temporary table exists as it
isn't populated in the schema and we don't want to give elevated privs
to run sys.table_exists().