Remove transition to move project back a round.
Research confirmed that the button on the project page to move a
project back a round was primarily to make it easier for siteadmins
to move projects that never had P2 back a round. Page states were
changed to the previous round's available state. This transition
has caused confusion, because it is not the same as a round repeat
but sounds like it should be.