Append newline to email footers
Previously, no final newline was included in the text emails.
When HTML emails were added, no final `<br>` was added.
This only appears to cause an issue where the email is send
to a mailing list, e.g. PPV mailing list. It appears that a
mailing list footer is appended, which probably includes
a newline followed by a line of underscores then some info
about the mailing list. This worked OK with text messages
since the newline put the underscores onto the line below,
but with the HTML version, a newline does nothing, and a
`<br>` is needed.

In fact, it seems reasonable that all our emails, text and
HTML, should have a terminating newline/`<br>`.