Only update upi.t_latest_home_visit for PP/PPVers
The only place in the code that uses upi.t_latest_home_visit is the
PP/PPV notification code, but every time any user even visits a project
page, an entry is added to the user_project_info table -- even if they
never proof a page or subscribe to an event.

Change it so that this field is only updated for current PP/PPVeres of
a project and add an upgrade script that will delete any rows that
don't satisfy this new requirement.