Only count PPers posted books in PPV report
Currently, not only posted books, but also those in
PPV Available, or being PPVed are counted and reported
in the PPV summary. This has the disadvantages that it
can jump up suddenly if a PPer uploads several projects
for PPV in a short time, and it can actually go downward
if a PPVer returns a project to the PPer for more work.
This can be confusing for the PPer/PPVer and maintainers
of the PPV records spreadsheet.

Changing to only count posted books will reduce the
occurrences of the first disadvantage and remove the
second completely.