Standardize "Return to x" links
As per task #1871, linked phrases such as
"Return to the project page" were inconsistently worded,
puncutated and marked up for translation.

Standardize return to project, round and Activity Hub.

Note that further standardization of capitalization
of "project page" could be done in future, where it
does not appear just as a simple "Return to" link.