Notify PM when project is checked out for PPing
There was already a notification if the project dropped
out of the rounds directly to an assigned PPer, and if
it went into the PP Available pool.

Task #1392 requested notification when it is checked
out of the pool or returned to the pool. The latter
has been implemented previously. This commit
implements the former.