Modify some Basic Latin picker sets
Some characters in the Basic Latin picker sets are confusing,
especially to new users who expect any character available in a
picker set to be preferred. Since the guidelines recommend not using
precomposed fractions, ordinals or superscripted numbers, this
commit removes them from the picker sets. It also removes the
stand-alone grave and acute accents, diaeresis, macron and cedilla,
as these cannot be used as combining diacriticals in Basic Latin.

The characters will remain in the Basic Latin character suite until
such time as there is a way to ensure that older projects which may
have used the characters and only have Basic Latin enabled will not
wind up with bad characters in the future.