Improve performance for neighbor stats
Not all users proof in all rounds, so the size of the users table
is always going to be bigger than the number of tallies in a round.
So instead of generating tallies for all users in a round and
trimming out those who aren't present, only get users with tallies
and special-case when the current user isn't in the tally round.
Performance is now based on how many users have proofread in a
round rather than the number of entries in the user table.