Include site structure definition in bootstrap.inc
How the site is structured, stages/rounds/pools, are integral to
virtually every page and the order matters. Create a file that
defines the structure (by including the structure-building files)
and include that once and only once in bootstrap.inc.

Furthermore, since every DP page must include base.inc,
site_vars.php is already loaded thanks to bootstrap.inc.

Update the includes checker to validate the best practices for which
files to include and not include.