api/v1: Add stats/site/projects/{stages,states} endpoints
Similar to the project counts on activity_hub.php.

Since it's two small DB queries for all the stats, and the output
is <700 bytes, don't bother filtering or grouping the stats by rounds.

NB because of the way Stages map to states, several states
are missing from the .../stages endpoint. These are:

* project_new
* proj_post_first_unavailable
* proj_post_complete
* proj_submit_pgposted
* project_complete
* project_delete