Fix set_from_clone()'s argument
set_from_clone() didn't accept a projectid, but we were passing it
one in one place but not the other. When a project is cloned, the
URL includes the projectid of the original project. But we also
include it in the POST as clone_projectid. Everything worked but it
did so in an unclear way.

Now use the clone_projectid value in the POST if we're getting data
from the POST, and from the GET param if we're servicing a clone
action.