SortUtility.inc: Remove 15 year old(!) bug comment and no-op code.
I think it's probably time to rip the bandaid off for this one.

The comment is 15 years old, from d1943a4ff07.

Since `$col` is undefined, the expression

`isset($this->_primaryValue) && @$col != $this->_primaryValue`

reduces to

`isset($this->_primaryValue) && null != $this->_primaryValue`

and the left and right operands of the `&&` are the same, so it
gives the same result if we just remove the RHS. Plus it makes
PHPstan happier.

For what it's worth, `$col` probably orginally came from iterating
over `$this->_collection` as done in `_parseSortingRule`