Run all tests on Ubuntu 20.04 and Ubuntu 22.04
Make sure we don't accidentally introduce breakages
on newer OSes.

NB the matrix definitions are duplicated across all the jobs for the moment.
To avoid this sharing we need to build a job that emits JSON as a single
line as a GITHUB_OUTPUT, and parse that using fromJSON in each job.
I struggled to get this right, and gave up for now.