Add pageload_smoketest.py and CI setup and jobs
A smoketest that sets up `site_vars.php` and a minimal set of database
tables so that PGDP can be served from the `php -S` CLI web server.

Then it loads many pages (soon to be every page) with representative
parameters and checks if there are any PHP errors, warnings, or
notices in the server logs.

Useful for, e.g., catching variables that aren't defined at runtime,
and other issues that PHPStan and the PHP linter can't find.

This is deliberately not an exhaustive test of all possible inputs --
smoke tests are meant to be fast and to weed out obvious brokenness.

That said, we probably should check each of, e.g. `task.php`'s
`action=` parameters so we get reasonable code coverage.

This is incomplete, but is good enough to be useful as is.