Patch erusev/parsedown to avoid a PHP Notice on malformed input
The notice is for indexing beyond the end of a string, and PHP's
result (an empty string) has logically the right effect in the
parser code, it just outputs an annoying notice which is visible
on the web page.

This is far from exhaustive. It would need to be done in a lot of
other places too, but serves as an interesting example of how to
patch vendored dependencies if we need to.