Disable autocapitalization of username input fields
The autocapitalize attribute is standard HTML5, but the default
behavior when it isn't explicitly set differs across browsers.

In particular we have some users using iPads, and in Safari,
autocapitalize is on by default.

https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/autocapitalize#usage_notes

Support goes back to iOS Safari 5 (2012); Chrome 43 (2015);
Opera 30 (2015); but not until Firefox 111 (2023). However the Firefox
non-support is harmless as it defaults to autocapitalize off anyway.