Revise the navbar
Introduce the idea of drop-down menu options to group like
things together and make better use of the top header space.
Use the new menu based on a user preference; if the preference
is set to "auto" (the default) dynamically use the menu based
on the width of the window.