Use xtext from vertilia/text to extract strings
xgettext has a nasty bug where it doesn't understand the PHP 7.3+
heredocs where the closing string is indented. This causes xgettext
to skip the entire rest of the document. Replace it with xtext.

Because we use xtext in prod, not just during development, we need
to require nikic/php-parser explicitly as a non-dev package.