Make "Available" tab mean "available to work in right now"
The original selection criteria for the "Available" tab was very
broad: the project was in some available state. That's not really
useful for most users who are thinking of "available" as available
for them to work on right now.

Change the tab to reflect this smaller set of available-to-them
projects and make it the first tab as its arguably the most useful now.