Adjust background job web proxy error detection
Until an HTTP stream returns some byte of data, the connection hasn't
succeeded. And until the stream connects we can't get any useful
metadata about it, such as if it timed out. This changes the web proxy
code to flush out the HTTP headers as quickly as possible while the
job runs, allowing us to better detect timeouts and handle other error
conditions.

As part of this, adjust the clean up job to more gracefully handle the
case where the TRASH folder doesn't exist, as the cleanup job will
delete it if the folder has no contents (and Remote File Manager will
create it again as-needed).