Fully type userprefs.php
The untyped functions in userprefs.php
were related to displaying individual
preference to the user.

The PR worked around the current logic
that invokes a specific function based on
the type of the preference to show the user
(e.g. a font selection dropdown or a textfield).
This means that we added some wide type for
now and may revisit this in the future.

Tested the PR by displaying and changing
some user prefs.