Type pinc/DPDatabase.inc
The function is core to all the code so this
PR adds some integration tests to make sure
that the basic operations still work.

The tests uncovered a few issues that were fixed,
like `close` wouldn't clear the connection or
`query` could return an unset variable.

The code in DPDatabase.pinc was switched to the
object-oriented API for easier typing but the
caller are mostly using the imperative API. The
mysqli documentation confirmed that this should
be fine, albeit ugly.