Type Activity/Stage/Round/Pool
Those 4 are closely coupled so it made
sense to handle them as a single PR.

Unfortunately due to the coupling, the return of
some key getters were not added, instead I added
a comment explaning the situation. Properly typing
will likely require some refactoring to split up
the per-type lists, which is better done in a follow-up.

TEST=Checked a user's available rounds, updated a project's state
to confirm that the rounds/pools were still listed.