Remove array_get
The function cannot be properly typed as it
is and fixing this would take a while, removing
it is shorter and better.

The change is pretty wide so care should be taken
as it is likely to introduce some regressions. I
have double-checked the diffs to make sure no
regression crept in but it is still possible.

TESTS=Did some PM/SA/P1 operations to confirm
that those still work.