[4/n] Remove array_get
The function cannot be properly typed as it
is and fixing this would take a while, removing
it is shorter and better.

The change updates accounts/*, locale/*,
project.php, tasks.php and userprefs.php.

TESTS=Created a new account, created a new task,
updated user prefs, and updated a project.