Upgrade PHPStan to 1.12.12
PHPStan was upgraded to 2.0 in November, but we're not ready
to upgrade as it detects a lot of new errors.
That said our first step is to upgrade to the latest 1.x.x
release, per:
https://github.com/phpstan/phpstan/blob/2.0.x/UPGRADING.md

There is one upside to this upgrade, which is to be able to
disable errors using their identifier (added in 1.11).

Also it seems like a low-value error was removed from L5,
so removed it from our configuration file.