[7/7] Remove array_get
The function cannot be properly typed as it
is and fixing this would take a while, removing
it is shorter and better.

The change removes usage in the last remaining
files. It also removes 2 tests for the function
that were converted instead of removed.

TESTS=Tested site/author searching and some
project updates.