Type pinc/misc.inc (#1386)
This exposed 2 issues:

1. In Page_MarkAsBad, `$reason` was typed as a string but it should
have been an int. The function and its callers were updated so we
have better typing moving forward.

2. `array_remove_invalid_keys` was not removing any keys if the list
of valid keys was empty. Fixed this as this is a potential security
issue (returning unexpected data in our API). Note that we ignore
`null` as it breaks existing cases. Added some unit tests for this.