Add JSON storage class & API access
Add a generic JSON-based storage class and an extension of it that
will be used in the API to store blobs from clients. This is a similar
structure to the Settings class / user_settings table but enforces
a JSON value.

This creates a shortcut in the API to allow JSON input to bypass
being deserialized and the output to bypass being serialized and
allow the routing function to manage that.