Handle edgecase of adding files to deleted projects
If a project has been deleted it doesn't show up in information_schema
and thus doesn't have a TABLE_COLLATION. If the user gets to add_files
with an old project page it would raise a deprecation warning. Also
prevent SAs getting to add_files if the project table doesn't exist.